+++ forEach.js
@@ -9,8 +9,12 @@
  * @param {function} callback The callback to apply. Params are: (value, key)
  */
 export default function forEach(object, callback) {
-	if(object === null || typeof object === 'undefined') return;
-	if(Array.isArray(object)) object.forEach(callback);
-	else if (typeof object.jquery !== 'undefined') object.each((k, v) => { callback($(v), k, v) });
-	else Object.entries(object).forEach(v => {callback(v[1], v[0])});
+	if (object === null || typeof object === 'undefined') return;
+	if (Array.isArray(object)) object.forEach(callback);
+	else if (typeof object.jquery !== 'undefined') object.each((k, v) => {
+		if (callback($(v), k, v) === false) return false;
+	});
+	else Object.entries(object).forEach(v => {
+			if (callback(v[1], v[0]) === false) return false;
+		});
 }